+++ stroom-task/stroom-task-impl/src/main/java/stroom/task/impl/TaskContextImpl.java
@@ -22,8 +22,13 @@ import stroom.task.api.TaskContext;
 import stroom.task.api.TaskTerminatedException;
 import stroom.task.api.TerminateHandler;
 import stroom.task.shared.TaskId;
+import stroom.util.NullSafe;
 
+import org.slf4j.Logger;
+
+import java.util.ArrayList;
 import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
@@ -63,6 +68,38 @@ public class TaskContextImpl implements TaskContext {
     @Override
     public void info(final Supplier<String> messageSupplier) {
         this.messageSupplier = messageSupplier;
+    }
+
+    @Override
+    public void info(final Supplier<String> messageSupplier, final Logger logger) {
+        this.messageSupplier = messageSupplier;
+
+        if (logger != null && logger.isDebugEnabled() && messageSupplier != null) {
+            logger.debug("Task stack: {}, user: {}, task: {}, info: {}",
+                    getTaskHierarchy(),
+                    NullSafe.get(userIdentity, UserIdentity::getId),
+                    name,
+                    messageSupplier.get());
+        }
+    }
+
+    private String getTaskHierarchy() {
+        if (taskId == null) {
+            return "";
+        } else {
+            final List<String> taskIds = new ArrayList<>();
+            TaskId currTaskId = this.taskId;
+            while (currTaskId != null) {
+                taskIds.add(currTaskId.getId());
+                currTaskId = currTaskId.getParentId();
+                // In case we get a huge hierarchy
+                if (currTaskId != null && taskIds.size() >= 10) {
+                    taskIds.add("...");
+                    break;
+                }
+            }
+            return String.join(" < ", taskIds);
+        }
     }
 
     @Override