+++ stroom-util/src/main/java/stroom/util/concurrent/UncheckedInterruptedException.java
@@ -1,6 +1,11 @@
 package stroom.util.concurrent;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 public class UncheckedInterruptedException extends RuntimeException {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(UncheckedInterruptedException.class);
 
     public UncheckedInterruptedException(final String message, final InterruptedException interruptedException) {
         super(message, interruptedException);
@@ -8,5 +13,19 @@ public class UncheckedInterruptedException extends RuntimeException {
 
     public UncheckedInterruptedException(final InterruptedException interruptedException) {
         super(interruptedException.getMessage(), interruptedException);
+    }
+
+//    public static void resetAndThrow(final InterruptedException e) throws UncheckedInterruptedException {
+//        LOGGER.debug(e.getMessage(), e);
+//        // Continue to interrupt the thread.
+//        Thread.currentThread().interrupt();
+//        throw new UncheckedInterruptedException(e);
+//    }
+
+    public static UncheckedInterruptedException create(final InterruptedException e) {
+        LOGGER.debug(e.getMessage(), e);
+        // Continue to interrupt the thread.
+        Thread.currentThread().interrupt();
+        return new UncheckedInterruptedException(e);
     }
 }