+++ systemInfo.sh
@@ -57,7 +57,7 @@ check_for_installed_binaries() {
 
 show_usage() {
   echo -e "A script to query Stroom's System Info providers on one or more hosts."
-  echo -e "Usage: ${BLUE}$0${GREEN} [-h] [-s] [-l hostList] [systemInfoName]${NC}"
+  echo -e "Usage: ${BLUE}$0${GREEN} [-h] [-s] [-l hostList] [systemInfoName] [[param_key=param_val]...]${NC}"
   echo -e "e.g:   ${BLUE}$0${GREEN}${NC}"
   echo -e "         # Select a provider from a list of names and query it on localhost.${NC}"
   echo -e "e.g:   ${BLUE}$0 stroom.dashboard.impl.ApplicationInstanceManager${NC}"
@@ -71,14 +71,16 @@ show_usage() {
   echo -e "${GREEN}-s${NC}:             Silent. Only output result. Useful for piping to other processes."
   echo -e "${GREEN}-l hostList${NC}:    A list of comma delimited hosts to connect to. Will use localhost if not provided."
   echo -e "${GREEN}systemInfoName${NC}: The name of the provider to query. If not provided you will be asked to select one."
+  echo -e "${GREEN}params${NC}:         Query parameter key=value paris, space delimited."
 }
 
 query_single_host() {
-  local host="$1"; shift
+  local host_port="$1"; shift
   local info_path="$1"; shift
-  local info_url="http://${host}:${port}${info_path}"
+  local info_url="http://${host_port}${info_path}"
 
-  debug_value "host" "${host}"
+
+  debug_value "host_port" "${host_port}"
   debug_value "info_url" "${info_url}"
   debug_value "http_auth_args" "${http_auth_args}"
 
@@ -87,25 +89,35 @@ query_single_host() {
       "at\n${BLUE}${info_url}${NC}"
   fi
 
+  local query_param_args=()
+  for param in "${params[@]}"; do
+    query_param_args+=( "--data-urlencode" "${param}" )
+  done
+
+  debug_value "query_param_args" "${query_param_args[*]}"
+
   # Call the api using curl
   local sys_info_json
   sys_info_json="$( \
     curl \
+      --get \
       --silent \
       --request GET \
       --header "${http_auth_args}" \
+      "${query_param_args[@]}" \
       "${info_url}" )"
 
   # Call the api but wrap the returned json inside a key that is the host name
   # Just output to stdout so the result can be piped to downstream processing,
   # e.g jq
-  jq "{ \"${host}\": .details }" \
+  jq "{ \"${host_port}\": .details }" \
     <<< "${sys_info_json}"
 }
 
 query_multiple_hosts() {
   local host_list="$1"; shift
   local info_path="$1"; shift
+  debug_value "host_list" "${host_list}"
   
   # TODO validate host_list to ensure it only contains [a-zA-Z.0-9\-] as
   # we are about to pass it to bash -c
@@ -119,6 +131,13 @@ query_multiple_hosts() {
   temp_dir="$(mktemp -d --suffix=_stroom_system_info)"
   debug_value "temp_dir" "${temp_dir}"
 
+  param_args_str=""
+  for param in "${params[@]}"; do
+    # shellcheck disable=SC2089
+    param_args_str+="\"${param}\" "
+  done
+  debug_value "param_args_str" "${param_args_str}"
+
   # Need to export these so they are visible in bash -c
   export temp_dir
   export http_auth_args
@@ -126,6 +145,8 @@ query_multiple_hosts() {
   export SCRIPT_DIR
   export SCRIPT_NAME
   export sys_info_name
+  # shellcheck disable=SC2090
+  export param_args_str
 
   # convert , to \n then use xargs to process each host in parallel
   # writing output to a temp file
@@ -139,11 +160,23 @@ query_multiple_hosts() {
     | xargs \
       -I'{}' \
       --max-procs="${max_proc_count}" \
-      bash -c 'echo "\"${SCRIPT_DIR}/${SCRIPT_NAME}\" -s -h \"{}\" \"${sys_info_name}\" > \"${temp_dir}/{}.json\""' \
+      bash -c 'echo "\"${SCRIPT_DIR}/${SCRIPT_NAME}\" -s -l \"{}\" \"${sys_info_name}\" ${param_args_str} > \"${temp_dir}/{}.json\""' \
     | xargs \
       -I'{}' \
       --max-procs="${max_proc_count}" \
       bash -c '{}'
+
+  if [ "${IS_DEBUG}" = true ]; then
+    for file in "${temp_dir}"/*.json; do
+      echo 
+      echo -e "${DGREY}DEBUG ${file} contents:${NC}"
+      echo -e "${DGREY}-START------------------------------------------------------${NC}"
+      local file_content
+      file_content="$(<"${file}")"
+      echo -e "${DGREY}${file_content}${NC}"
+      echo -e "${DGREY}-END--------------------------------------------------------${NC}"
+    done
+  fi
 
   # Merge all the files into one json object, which each host's content as a top
   # level key. This is output to stdout for onward piping if needs be
@@ -158,7 +191,7 @@ query_multiple_hosts() {
 # containts safe chars
 validate_host_list() {
   local host_list="$1"; shift
-  local host_list_pattern="^[-.,0-9a-zA-Z]+$"
+  local host_list_pattern="^([-.0-9a-zA-Z]+(:[0-9]+)?,?)+$"
   if [[ ! "${host_list}" =~ ${host_list_pattern} ]]; then
     error_exit "Host list [${host_list}] does not match pattern ${host_list_pattern}"
   fi
@@ -176,7 +209,60 @@ validate_sys_info_name() {
   fi
 }
 
-main(){
+capture_params() {
+  local sys_info_name="$1"; shift
+  local params_url="${params_url_base}/${sys_info_name}"
+
+  debug "Querying params for ${sys_info_name}"
+  param_details="$( \
+    curl \
+        --silent \
+        --request GET \
+        --header "${http_auth_args}" \
+        "${params_url}" \
+      | jq -r '. | sort_by(.paramType, .name)[] | .name + " (" + .paramType + ") - " + .description' \
+    )"
+
+  if [[ -n "${param_details}" ]]; then
+    debug "Found some params"
+
+    # Get just the keys
+    local param_keys
+    param_keys="$( \
+      curl \
+          --silent \
+          --request GET \
+          --header "${http_auth_args}" \
+          "${params_url}" \
+          | jq -r '.[] | .name' )"
+    param_keys_arr=()
+    while IFS= read -r line; do
+      param_keys_arr+=("$line")
+    done <<< "${param_keys}"
+
+    local mandatory_param_details
+    mandatory_param_details="$( \
+      grep "MANDATORY" <<<"${param_details}" )"
+
+    if [[ -n "${mandatory_param_details}" ]]; then
+      echo -e "${GREEN}Mandatory query parametes are required for this system info provider${NC}"
+      echo -e "${YELLOW}${param_details}${NC}"
+
+      for param_key in "${param_keys_arr[@]}"; do
+        echo -n "Enter the value for parameter ${param_key} (or hit enter for no value): "
+        read -r param_value
+        debug_value "param_value" "${param_value}"
+        if [[ -n "${param_value}" ]]; then
+          params+=( "${param_key}=${param_value}" )
+        fi
+      done
+
+      debug_value "params" "${params[*]}"
+    fi
+  fi
+}
+
+main() {
   check_for_installed_binaries
 
   #optspec="gu:rih:"
@@ -216,7 +302,17 @@ main(){
   shift $((OPTIND -1))
   #echo "Remaining args [${@}]"
 
-  local name_arg="$1"
+  local name_arg
+  if [ $# -gt 0 ]; then
+    name_arg="$1"; shift
+  fi
+
+  local params=()
+  #loop through the pairs of args
+  while [ $# -gt 0 ]; do
+    params+=( "$1" )
+    shift
+  done
 
   SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
   SCRIPT_NAME="$0"
@@ -227,6 +323,7 @@ main(){
   debug_value "name_arg" "${name_arg}"
   debug_value "SCRIPT_DIR" "${SCRIPT_DIR}"
   debug_value "SCRIPT_NAME" "${SCRIPT_NAME}"
+  debug_value "params" "${params[*]}"
 
   validate_host_list "${host_list}"
 
@@ -245,18 +342,37 @@ main(){
     local api_token='eyJhbGciOiJSUzI1NiIsImtpZCI6IjhhM2I1OGNhLTk2ZTctNGFhNC05ZjA3LTQ0MDBhYWVkMTQ3MSJ9.eyJleHAiOjE2Njg3NzczNjcsInN1YiI6ImFkbWluIiwiaXNzIjoic3Ryb29tIiwiYXVkIjoiTlhDbXJyTGpQR2VBMVN4NWNEZkF6OUV2ODdXaTNucHRUbzZSdzVmTC5jbGllbnQtaWQuYXBwcy5zdHJvb20taWRwIn0.YhhRQKF29CQm7IWHP2sh-i70qBicuWKJLhH5UmSRxeyHJ2T38RmHVEcIjC9tv71-gJR4z3bY9tRq_r6cf5hG2G9DKVfPoVZTN-MhK-pU5eD3VbVc2HBEm0Xk02LL7vKRS2mKLolaI-DC_5TuYclZ-CGKmmqhh8Bb5evqAXnknccILAGsl3xDFzsKdfuX5iZ5wjizxvgyMjvLCfaM6P-Ut5kUWYHmxpdNextE3p35Kajw_iEbHvZ3_CobX09l5QQo2R7Hgzk3lIFrGxUqQ7jyC_DmmxRlcT-pyJApl7_TFlFJm153R_9gDXIwphI1mg1vsAojkWncCl8ODY227t3-Lw'
   fi
 
-  local port=8080
-  local default_host="localhost"
+  local default_port="8080"
+  local default_host_port="localhost:8080"
   local base_path="/api/systemInfo/v1"
-  local names_path="${base_path}/names"
+  local names_path_part="/names"
+  local params_path_part="/params"
   local http_auth_args="Authorization:Bearer ${api_token}"
-  local base_url="http://localhost:8080/api/systemInfo/v1"
-  local names_url="${base_url}/names"
 
   if [[ "${is_silent}" = false ]]; then
     echo -e "${GREEN}Using names URL ${BLUE}${names_url}${NC}"
     echo
   fi
+
+  local host
+  if [[ -n "${host_list}" ]]; then
+    # Get the first host in the list
+    host="${host_list%%,*}"
+  else
+    host="${default_host}"
+  fi
+  local host_port="${host}"
+  if [[ ! "${host}" =~ :[0-9]+ ]]; then
+    host_port="${host}:${default_port}"
+  fi
+
+  local names_url="http://${host_port}${names_path_part}"
+  debug_value "names_url" "${names_url}"
+  local base_url="http://${host_port}${base_path}"
+  local names_url="${base_url}${names_path_part}"
+  local params_url_base="${base_url}${params_path_part}"
+  debug_value "names_url" "${names_url}"
+  debug_value "params_url_base" "${params_url_base}"
 
   local sys_info_name
 
@@ -264,17 +380,7 @@ main(){
     sys_info_name="${name_arg}"
   else
     # Use fzf to get a system info name
-    local host
-    if [[ -n "${host_list}" ]]; then
-      # Get the first host in the list
-      host="${host_list%%,*}"
-    else
-      host="${default_host}"
-    fi
-
-    local names_url="http://${host}:${port}${names_path}"
-    debug_value "names_url" "${names_url}"
-
+    debug "Qurying for system info provider names"
     local sys_info_names_json
     sys_info_names_json="$( \
       curl \
@@ -282,6 +388,8 @@ main(){
         --request GET \
         --header "${http_auth_args}" \
         "${names_url}" )"
+
+    debug_value "sys_info_names_json" "${sys_info_names_json}"
 
     # Extract a list of names from the json
     sys_info_names="$( \
@@ -329,6 +437,13 @@ main(){
 
   validate_sys_info_name "${sys_info_name}"
 
+  local param_details=()
+  if [[ "${is_silent}" = false ]] && [[ "${#params[@]}" -eq 0 ]]; then
+    # No param args used, so see if the sys info provider has any params
+    # and if so prompt user for their values.
+    capture_params "${sys_info_name}"
+  fi
+
   # Construct the url path base on the sys info provider name
   local info_path="${base_path}/${sys_info_name}"
 
@@ -338,9 +453,9 @@ main(){
   else
     local host="${host_list:-${default_host}}"
     debug "Single host"
-    debug_value "host" "${host}"
+    debug_value "host_port" "${host_port}"
 
-    query_single_host "${host}" "${info_path}"
+    query_single_host "${host_port}" "${info_path}"
   fi
 }